% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustVertex.R
\name{adjustVertex}
\alias{adjustVertex}
\title{Modify given coordinates of given vertices of a simplex}
\usage{
adjustVertex(simplex, newcoords, overwrite = FALSE)
}
\arguments{
\item{simplex}{\code{'chsmplx'} type object containig most information
of simplex.}

\item{newcoords}{List with elements named like the vertices to be modified.
Each element must have a vector with the (ordered) coordinates
setted for the experiment. containig a dataframe in which the variables
to be modified to that vertex are provided}

\item{overwrite}{logical argument indicating if the simplex must be
automatically overwrited}
}
\value{
A 'chsmplx' type object with the information of the simplex
       including the conditions for the new experiment to be permormed.
}
\description{
Changes the coordinates of generated vertices when slightly differences
were impossible to avoid at the moment of setting the experiment
(e.g. small differences is mass components when preparing a mixture).
}
