% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSurfaces.R
\name{exampleSurfaceR2}
\alias{exampleSurfaceR2}
\title{Evaluates two variables in a hypothetical response surface}
\usage{
exampleSurfaceR2(temp, pH, noise = 0)
}
\arguments{
\item{temp}{Numeric value (may be a vector) for temperature.}

\item{pH}{Numeric value (may be a vector) for pH.}

\item{noise}{Noise used in the response surface as absolute percentaje.}

\item{seed}{Seed for the noise generated}
}
\value{
The yield of the reaction
}
\description{
The response surface modelates the yield of a hypotetical chemical reaction
affected by the temperature and the pH of the media.
}
\examples{
exampleSurfaceR2(temp = 290, pH = 4)
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Jesús Ágreda, \email{jagreda@unal.edu.co}
}
