% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmplsurface.R
\name{exmplsurface}
\alias{exmplsurface}
\title{Evaluates two variables in a hypothetical response surface}
\usage{
exmplsurface(temp, pH, noise = 0, seed = NULL)
}
\arguments{
\item{temp}{Numeric value (may be a vector) for temperature.}

\item{pH}{Numeric value (may be a vector) for pH.}

\item{noise}{Noise used in the response surface as absolute percentaje.}

\item{seed}{Seed for the noise generated}
}
\value{
The yield of the reaction
}
\description{
The response surface modelates the yield of a hypotetical chemical reaction
affected by the temperature and the pH of the media.
}
\examples{
exmplsurface(temp = 290, pH = 4)
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Jesús Ágreda, \email{jagreda@unal.edu.co}
}
