% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsimplex-package.R
\docType{package}
\name{labsimplex-package}
\alias{labsimplex-package}
\title{\code{labsimplex}: Simplex Optimization Algorithms for Laboratory and
Processes Applications}
\description{
The labsimplex package provides the Implementation of the simplex
optimization algorithms firstly proposed by Spendley et al. (1962)
<doi:10.1080/00401706.1962.10490033> and later modified by Nelder
and Mead (1965) <doi:10.1093/comjnl/7.4.308> for laboratory and
processes applications. Tools for simplexes' graphical representation
and some example surface responses are provided.
}
\details{
The package uses list objects of class \code{'smplx'} to store the
simplex information including all the vertices coordinates and experimental
responses.
}
\section{labsimplex functions}{

The labsimplex functions can generate a new \code{'smplx'} class object,
assing responses to the vertices to generate the next one and to visualize
different spatial representation of the \emph{n}-dimensional simplex in 2D
or 3D proyections. Detailed information can be found by typing
\code{vignette('labsimplex')}.
}

\references{
Nelder, J. A., and R. Mead. 1965. “A Simplex Method for Function
Minimization.” The Computer Journal 7 (4): 308–13.

Spendley, W., G. R. Hext, and F. R0. Himsworth. 1962. “Sequential
Application of Simplex Designs in Optimization and Evolutionary Operation.”
Technometrics 4 (4): 441–61.
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Jesús Ágreda, \email{jagreda@unal.edu.co}
}
