% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/labsimplex.R
\docType{package}
\name{labsimplex}
\alias{labsimplex}
\alias{labsimplex-package}
\title{labsimplex: Implementation of simplex optimization algorithms for
laboratory applications.}
\usage{
labsimplex(N = NULL, start = NULL, centroid = NULL,
  stepsize = NULL, neg.con = FALSE, zer.con = FALSE,
  min.val = NULL, usrdef = NULL, var.name = NULL)
}
\arguments{
\item{N}{number of dimentions in the space.}

\item{start}{numeric vector of size \code{N} with initial coordinates for the first vertex.}

\item{centroid}{coordinates of centroid for initial simplex.}

\item{stepsize}{numeric vector of size \code{N} with the step size for each coordinate.}

\item{neg.con}{logical argument indicating if there is a constraing for only non-negative coordinates.}

\item{zer.con}{logical argument indicating if there is a constraing for zero values in the coordinates.}

\item{min.val}{numeric vector of size \code{N} containing the minimal that a variable can
approach zero if \code{zer.con = TRUE}.}

\item{usrdef}{\code{(N+1)xN} matrix containig in (N+1) rows the N coordinates for each vertex.}

\item{var.name}{vector containing the names for the variables.}
}
\value{
A \code{smplx} type object with the information of the simplex.
}
\description{
Generates a \code{smplx} class object containing the coordinates
of the N+1 vertexes of a simplex in a N-dimentional space.
Addionally can check if faces for given coordinates of a used defined
simplex are in different hyperplanes.
}
\details{
All parameters are optional. If no parameters are provided a void
\code{smplx} class object is generated. Once the simplex is genereated,
the experiments under the conditions indicated for each variable at each
vertex must be carried and the quality function evaluated. Those quality
function values are assigned to the \code{smplx} object at the moment of
generating the new vertex (see \code{\link{generateVertex}}). The initial simplex
coordinates are generated following the
\href{https://bit.ly/2u8HlYt}{general algorithm} for the cartesian
coordinates for regular n-dimensional simplex.
}
\section{labsimplex functions}{

The labsimplex functions...
}

\examples{
  labsimplex(N = 3)
  labsimplex(N = 3, centroid = c(66, 1, 12), stepsize = c(10, 0.1, 2),
             neg.con = TRUE, var.name = c('potential', 'pH', 'T'))

  labsimplex(usrdef = rbind(c(2, 0, 0), c(-0.5, 1, 0), c(-0.5, -0.7, 1),
                            c(-0.5, -0.4, -0.6)))
\dontrun{
## A user defined coordinates may define faces that rely on same hyperplane:
  labsimplex(usrdef = rbind(c(2, 0, 0), c(-0.5, 0, 0), c(0, 0, 0), c(-0.5, -0.4, -0.6)))
}
}
