% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimplexResponse.R
\name{plotSimplexResponse}
\alias{plotSimplexResponse}
\title{Plots the response versus the vertex number of a simplex optimization.}
\usage{
plotSimplexResponse(x, ...)
}
\arguments{
\item{x}{\code{smplx} class object containig the coordinates of
the vertices and their response values.}

\item{...}{other graphical parameters used in plot()}
}
\value{
Plot of quiality function against vertex number
}
\description{
The function generates a plot for a \code{smplx} class object whose vertices
must have a response value assigned. The response is plotted
against the vertex number. The only vertex allowed to not having response
assigned is the last one.
}
\details{
If the simplex object being ploted was obtained using a variable size
algorithm, some experimental points could be disregarded and will be
shown with a red mark indicating that the vertex was not used in
new vertices calculations. Those points are also ignored by the line that
links the data in the scaterplot and pretends to show a tendency to
better values as more vertices are evaluated.
}
\examples{
set.seed(1)
# Generate a simplex and add some responses to the vertices
simplex <- labsimplex(N = 3)
generateVertex(simplex, qflv = rnorm(4), algor = 'variable',
  overwrite = TRUE)
generateVertex(simplex, qflv = 2.75, algor = 'variable', overwrite = TRUE)
generateVertex(simplex, qflv = 0.8,  algor = 'variable', overwrite = TRUE)
generateVertex(simplex, qflv = 2.94, algor = 'variable', overwrite = TRUE)
generateVertex(simplex, qflv = 3.14, algor = 'variable', overwrite = TRUE)

# Plot the response versus the vertex function
plotSimplexResponse(simplex)
}
