% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.smplx.R
\name{print.smplx}
\alias{print.smplx}
\title{S3 method \code{print} for simplex objects}
\usage{
\method{print}{smplx}(x, extended = FALSE, conventions = TRUE, ...)
}
\arguments{
\item{x}{simplex (object of class to be printed}

\item{extended}{logical, if \code{TRUE}, the object is printed as a
list containing all hiden elements}

\item{conventions}{logical, if \code{TRUE} (default), the conventions
used are printed}

\item{...}{more parameters passed to \code{\link[base]{print}}}
}
\description{
Prints simplex information.
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Jesús Ágreda, \email{jagreda@unal.edu.co}
}
