% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexExport.R
\name{simplexExport}
\alias{simplexExport}
\title{Exports the information contained on a \code{smplx} class object.}
\usage{
simplexExport(simplex, filename = NULL, direc = NULL)
}
\arguments{
\item{simplex}{\code{smplx} class object containing the simplex to be
stored.}

\item{filename}{string with the name (without extention) of the file that
will be created. If not provided, the name of the simplex
object will be used.}

\item{direc}{directory in which the file will be saved. If not provided,
the current directory will be used.}
}
\value{
A \code{.smplx} file containing all the information required to
        continue with the optimization process once all vertex experiments
        have been carried.
}
\description{
Creates a \code{.smplx} file (text file) that stores the information
contained in the simplex (a \code{smplx} class object, see
\code{\link{labsimplex}}). It allows to continue with the optimization
proccess after experiment described by a vertex has been carried and
a quality function value has been obtained.
}
\examples{
\dontrun{
  simplex <- labsimplex(N = 5, qual.fun = rnorm(6, 2, 1))
  simplexExport(simplex = simplex)
}
}
\seealso{
\code{\link{simplexImport}}
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Jesús Ágreda, \email{jagreda@unal.edu.co}
}
