% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexImport.R
\name{simplexImport}
\alias{simplexImport}
\title{Imports the information contained in a \code{.smplx} file.}
\usage{
simplexImport(filename, auth.ld = TRUE, name = NULL)
}
\arguments{
\item{filename}{string with the name of the file (with extension) to
be imported. The file must be generated using
\code{\link{simplexExport}}. It may include path if the
file is not in the current directory.}

\item{auth.ld}{logical. Should the imported simplex object be directly
loaded on the Environment? Default to TRUE.}

\item{name}{name for the simplex object to be created if
\code{auth.ld = FALSE}. When not provided, the ID in the
file will be used.}
}
\value{
A \code{smplx} class object with the complete information of the
        simplex
}
\description{
Imports the information contained in a \code{.smplx} file and creates a
\code{'smplx'} class object.
}
\examples{
\dontrun{
  simplexExport(labsimplex(N = 4), name = "simplex4D")
  simplexImport("simplex4D.smplx")
}
}
