% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleOptim.R
\name{addSimplex2Surface}
\alias{addSimplex2Surface}
\title{Adds simplex movements to a response surface contour}
\usage{
addSimplex2Surface(p, simplex)
}
\arguments{
\item{p}{asdsdf}

\item{simplex}{simplex}
}
\description{
The function complements the
}
\examples{
simplex <- exampleOptimization(surface = exampleSurfaceR2,
                               centroid = c(7, 340),
                               stepsize = c(1.2, 15))
(p <- cntr(surface = exampleSurfaceR2))
p <- addSimplex2Surface(p = p, simplex = simplex)
print(p)
}
\seealso{
\code{\link{cntr}} \code{\link{exampleOptimization}}
}
