% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustVertex.R
\name{adjustVertex}
\alias{adjustVertex}
\title{Modify given coordinates of given vertices of a simplex}
\usage{
adjustVertex(simplex, newcoords, overwrite = FALSE)
}
\arguments{
\item{simplex}{\code{'chsmplx'} type object containig most information
of simplex.}

\item{newcoords}{List with elements named like the vertices to be modified.
Each element must have a vector with the actual (ordered)
coordinates used in the experiment. \code{NA} may be used
to indicate coordinates that were unchanged.}

\item{overwrite}{logical argument. If \code{TRUE} the output simplex will
replace the provided one in the \code{simplex} parameter.
Defauklt \code{overwrite = TRUE}}
}
\value{
A 'chsmplx' type object with the information of the simplex
       including the conditions for the new experiment to be permormed.
}
\description{
Changes the coordinates of generated vertices when slightly differences
were impossible to avoid at the moment of setting the experiment
(e.g. small differences in mass components when preparing a mixture).
}
\examples{
simplex <- labsimplex(N = 3, start = c(7, 25, 0.15),
                      stepsize = c(0.2, 5, 0.02))
adjustVertex(simplex = simplex, newcoords = list(Vertex.1 = c(7, NA, NA),
                                                 Vertex.3 = c(7.2, NA, NA)),
             overwrite = TRUE)
}
