% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleOptim.R
\name{exampleOptimization}
\alias{exampleOptimization}
\title{Performs a complete simplex optimization}
\usage{
exampleOptimization(surface, simplex = NULL, centroid = c(7, 3.4),
  stepsize = c(0.6, 10), algor = "fixed", experiments = 17, noise = 0)
}
\arguments{
\item{surface}{example response surface to be used. See
\code{\link{exampleSurfaceR2}} and \code{\link{exampleSurfaceR2.2pks}}.}

\item{simplex}{\code{'smplx'} type object containig simplex information}

\item{centroid}{coordinates of centroid for initial simplex}

\item{stepsize}{numeric vector of size \code{N} with the step size for each
coordinate}

\item{algor}{algorithm to be followed in the vertex generation.
\code{"fixed"} for a fixed-size simplex following
Spendley (1962) algorithm or \code{"variable"} for a
variable-size simplex following Nelder and Mead (1965)
algorithm}

\item{experiments}{number of vertexes to evaluate}

\item{noise}{absolute noise to be included in the results}
}
\description{
The function plots a perspective 3D plot for the example response surfaces included in the package
}
