% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.smplx.R
\name{plot.smplx}
\alias{plot.smplx}
\title{Makes a two dimentional plot of a simplex object.}
\usage{
\method{plot}{smplx}(x, sel.dim = NULL, all.ver = TRUE,
  all.lin = TRUE, expand = TRUE, exp.fac = 1.5, ...)
}
\arguments{
\item{x}{\code{smplx} class object containig the coordinates of
the vertices.}

\item{sel.dim}{numeric or char vector for variables to be considered when
simplex dimensionality is higher than 2. If \code{numeric}
form it must contain dimensions ordinal number. If
\code{char}, it must contain dimensions names.}

\item{all.ver}{logical. Should all vertex be plotted? If \code{FALSE}
draws only vertices corresponding to current simplex.}

\item{all.lin}{logical. Should all lines be drawn? If \code{FALSE} draws
only last simplex.}

\item{expand}{logical. Should the plot scales be expanded?}

\item{exp.fac}{expansion factor used when \code{expand = TRUE}.}

\item{...}{other graphical parameters used in plot()}
}
\value{
2D proyection of the simplex coordinates
}
\description{
The function generates a 2D plot for a \code{smplx} class object having
dimensionality of at least 2. When dimensionality is higher than 2, the
function plots a 2D proyection of selected dimensions.
}
\details{
For 3D representations of simplex with dimensionality higher than 2 use
\code{\link{plotSimplex3D}}
}
\examples{
  plot(x = labsimplex(N = 2))
  plot(x = labsimplex(N = 2), expand = FALSE)

  plot(x = labsimplex(N = 8))
  plot(x = labsimplex(N = 8), sel.dim = c(3, 4))

  ## Simulation of the real proccess where a simplex is made and QF measured,
  set.seed(12)
  simplex2D <- labsimplex(N = 2)
  plot(x = simplex2D)
  generateVertex(simplex = simplex2D, qflv = rnorm(3), overwrite = TRUE)
  plot(x = simplex2D)
  generateVertex(simplex = simplex2D, qflv = rnorm(1), overwrite = TRUE)
  plot(x = simplex2D)
  generateVertex(simplex = simplex2D, qflv = rnorm(1), overwrite = TRUE)
  plot(x = simplex2D)
}
