% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.smplx.R
\name{print.smplx}
\alias{print.smplx}
\title{Prints given simplex (a \code{smplx} class object)}
\usage{
\method{print}{smplx}(x, extended = FALSE, conventions = TRUE, ...)
}
\arguments{
\item{x}{simplex object to be printed}

\item{extended}{logical, if \code{TRUE}, the object is printed as a list containing all elements}

\item{conventions}{logical, if \code{TRUE} (default), the conventions used are printed}

\item{...}{other arguments passed to print}
}
\description{
Prints given simplex information.
}
